Refactor kvm functions for snphost ok
